+++ src/main/java/tconstruct/util/config/PHConstruct.java
@@ -48,12 +48,12 @@ public class PHConstruct
         miningLevelIncrease = config.get("Difficulty Changes", "Modifiers increase Mining Level", true).getBoolean(true);
         denyMattock = config.get("Difficulty Changes", "Deny creation of non-metal mattocks", false).getBoolean(false);
 		
-		ingotsPerOre = config.get("Smeltery Output Modification", "Ingots per ore", 2, "Number of ingots returned from smelting ores in the smeltery").getInt(2);
-		ingotsBronzeAlloy = config.get("Smeltery Output Modification", "Bronze ingot return", 3, "Number of ingots returned from smelting Bronze in the smeltery").getInt(3);
-		ingotsAluminumBrassAlloy = config.get("Smeltery Output Modification", "Aluminum Brass ingot return", 4, "Number of ingots returned from smelting Aluminum Brass in the smeltery").getInt(4);
-		ingotsAlumiteAlloy = config.get("Smeltery Output Modification", "Alumite ingot return", 3, "Number of ingots returned from smelting Alumite in the smeltery").getInt(3);
-		ingotsManyullynAlloy = config.get("Smeltery Output Modification", "Manyullyn ingot return", 1, "Number of ingots returned from smelting Manyullyn in the smeltery").getInt(1);
-		ingotsPigironAlloy = config.get("Smeltery Output Modification", "Pig Iron ingot return", 1, "Number of ingots returned from smelting Pig Iron in the smeltery").getInt(1);
+		ingotsPerOre = config.get("Smeltery Output Modification", "Ingots per ore", 2, "Number of ingots returned from smelting ores in the smeltery").getDouble(2);
+		ingotsBronzeAlloy = config.get("Smeltery Output Modification", "Bronze ingot return", 3, "Number of ingots returned from smelting Bronze in the smeltery").getDouble(3);
+		ingotsAluminumBrassAlloy = config.get("Smeltery Output Modification", "Aluminum Brass ingot return", 4, "Number of ingots returned from smelting Aluminum Brass in the smeltery").getDouble(4);
+		ingotsAlumiteAlloy = config.get("Smeltery Output Modification", "Alumite ingot return", 3, "Number of ingots returned from smelting Alumite in the smeltery").getDouble(3);
+		ingotsManyullynAlloy = config.get("Smeltery Output Modification", "Manyullyn ingot return", 1, "Number of ingots returned from smelting Manyullyn in the smeltery").getDouble(1);
+		ingotsPigironAlloy = config.get("Smeltery Output Modification", "Pig Iron ingot return", 1, "Number of ingots returned from smelting Pig Iron in the smeltery").getDouble(1);
 		
         //1467-1489
         woodStation = config.getBlock("Wood Tool Station", 1471).getInt(1471);
@@ -677,12 +677,12 @@ public class PHConstruct
     public static boolean denyMattock;
 
 	//Smeltery Output Modification
-	public static int ingotsPerOre;
-	public static int ingotsBronzeAlloy;
-	public static int ingotsAluminumBrassAlloy;
-	public static int ingotsAlumiteAlloy;
-	public static int ingotsManyullynAlloy;
-    public static int ingotsPigironAlloy;
+	public static double ingotsPerOre;
+	public static double ingotsBronzeAlloy;
+	public static double ingotsAluminumBrassAlloy;
+	public static double ingotsAlumiteAlloy;
+	public static double ingotsManyullynAlloy;
+    public static double ingotsPigironAlloy;
 
     //Ultra Hardcore modifiers
     public static boolean enableHealthRegen;
